fix: 리뷰 생성 시 참조하는 id가 존재하지 않을 때 예외 처리

+++ src/main/java/com/example/jomajomadelivery/review/service/ReviewService.java
@@ -1,12 +1,16 @@
 package com.example.jomajomadelivery.review.service;
 
+import com.example.jomajomadelivery.exception.CustomException;
 import com.example.jomajomadelivery.review.dto.request.ReviewCreateRequestDto;
 import com.example.jomajomadelivery.review.dto.request.ReviewUpdateRequestDto;
 import com.example.jomajomadelivery.review.dto.response.ReviewResponseDto;
 import com.example.jomajomadelivery.review.entity.Review;
+import com.example.jomajomadelivery.review.exception.ReviewErrorCode;
 import com.example.jomajomadelivery.review.repository.ReviewRepository;
 import com.example.jomajomadelivery.store.entity.Store;
 import com.example.jomajomadelivery.store.repository.StoreRepository;
+import com.example.jomajomadelivery.user.entity.User;
+import com.example.jomajomadelivery.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,19 +23,38 @@ import org.springframework.web.server.ResponseStatusException;
 @RequiredArgsConstructor
 public class ReviewService {
     private final ReviewRepository reviewRepository;
-//    private final UserRepository userRepository;
+    private final UserRepository userRepository;
     private final StoreRepository storeRepository;
 
 
     @Transactional
     public ReviewResponseDto create(Long storeId, ReviewCreateRequestDto dto) {
-//        User foundUser = userRepository.getReferenceById(dto.uesrId());
-        // 현재 로그인된 유저 -> 리뷰 작성자
-        // FIXME:
-        Store foundStore = storeRepository.getReferenceById(storeId);
+        if (dto.rating()<1 || 5<dto.rating()) {
+            throw new CustomException(ReviewErrorCode.RATING_NOT_VALID);
+        }
 
-//        Review review = Review.createReview(foundUser, foundStore, dto);
-        Review review = Review.createReview(1L, foundStore , dto);
+        if (dto.contents().length()>200) {
+            throw new CustomException(ReviewErrorCode.CONTENTS_LENGTH_NOT_VALID);
+        }
+
+        // 현재 로그인된 유저 -> 리뷰 작성자
+        // FIXME: 실제 로그인된 userId값으로 대체
+        Long userId = 1L;
+
+        // 로그인된 사용자로 부터 가져온 ID인데 무슨 경우?
+        if (!userRepository.existsById(userId)) {
+            // FIXME: user 쪽 예외로 대체
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No such user id.");
+        }
+        User userRef = userRepository.getReferenceById(storeId);
+
+        if (!storeRepository.existsById(storeId)) {
+            // FIXME: store 쪽 예외로 대체
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No such store id.");
+        }
+        Store storeRef = storeRepository.getReferenceById(storeId);
+
+        Review review = Review.createReview(userRef, storeRef , dto);
         Review savedReview = reviewRepository.save(review);
 
         return ReviewResponseDto.toDto(savedReview);
@@ -49,6 +72,14 @@ public class ReviewService {
 
     @Transactional
     public ReviewResponseDto update(Long id, ReviewUpdateRequestDto dto) {
+        if (dto.rating()<1 || 5<dto.rating()) {
+            throw new CustomException(ReviewErrorCode.RATING_NOT_VALID);
+        }
+
+        if (dto.contents().length()>200) {
+            throw new CustomException(ReviewErrorCode.CONTENTS_LENGTH_NOT_VALID);
+        }
+
         Review foundReview = reviewRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No such review id: " + id));
         foundReview.update(dto);
         return ReviewResponseDto.toDto(foundReview);
@@ -56,7 +87,9 @@ public class ReviewService {
 
     @Transactional
     public void delete(Long id) {
-        // TODO: not found 에러 처리?
+        if (!reviewRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No such review id: " + id);
+        }
         reviewRepository.deleteById(id);
     }
 }
